+++ src/status.rs
@@ -16,6 +16,8 @@ pub struct Location {
     pub address: Option<String>,
     pub lat: f64,
     pub lon: f64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub timezone: Option<String>,
 }
 
 #[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
@@ -190,6 +192,41 @@ pub struct Stream {
     pub ustream: Option<String>,
 }
 
+#[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
+pub struct Link {
+    pub name: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+    pub url: String,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+#[serde(rename_all = "snake_case")]
+pub enum BillingInterval {
+    Yearly,
+    Monthly,
+    Weekly,
+    Daily,
+    Hourly,
+    Other,
+}
+
+impl Default for BillingInterval {
+    fn default() -> Self {
+        BillingInterval::Monthly
+    }
+}
+
+#[derive(Serialize, Deserialize, Default, Debug, Clone, PartialEq)]
+pub struct MembershipPlan {
+    pub name: String,
+    pub value: f64,
+    pub currency: String,
+    pub billing_interval: BillingInterval,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub description: Option<String>,
+}
+
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 pub enum ApiVersion {
     #[serde(rename = "14")]
@@ -227,6 +264,10 @@ pub struct Status {
     pub events: Option<Vec<Event>>,
     #[serde(skip_serializing_if = "Option::is_none")]
     pub radio_show: Option<Vec<RadioShow>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub links: Option<Vec<Link>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub membership_plans: Option<Vec<MembershipPlan>>,
 
     // SpaceAPI internal usage
     #[serde(skip_serializing_if = "Option::is_none")]
@@ -300,6 +341,8 @@ pub struct StatusBuilder {
     feeds: Option<Feeds>,
     events: Option<Vec<Event>>,
     radio_show: Option<Vec<RadioShow>>,
+    links: Option<Vec<Link>>,
+    membership_plans: Option<Vec<MembershipPlan>>,
     issue_report_channels: Vec<IssueReportChannel>,
     extensions: Extensions,
     state: Option<State>,
@@ -387,6 +430,16 @@ impl StatusBuilder {
         self
     }
 
+    pub fn add_link(mut self, link: Link) -> Self {
+        self.links.get_or_insert(vec![]).push(link);
+        self
+    }
+
+    pub fn add_membership_plan(mut self, membership_plan: MembershipPlan) -> Self {
+        self.membership_plans.get_or_insert(vec![]).push(membership_plan);
+        self
+    }
+
     pub fn add_project<S: Into<String>>(mut self, project: S) -> Self {
         self.projects.get_or_insert(vec![]).push(project.into());
         self
@@ -448,6 +501,17 @@ impl StatusBuilder {
             if self.state.is_none() {
                 return Err("state must be present in v0.13".into());
             }
+            if let Some(ref location) = self.location {
+                if location.timezone.is_some() {
+                    return Err("location.timezone is only present in v0.14 and above".into());
+                }
+            }
+            if self.links.is_some() {
+                return Err("links is only present in v0.14 and above".into());
+            }
+            if self.membership_plans.is_some() {
+                return Err("membership_plans is only present in v0.14 and above".into());
+            }
         }
 
         Ok(Status {
@@ -464,6 +528,8 @@ impl StatusBuilder {
             feeds: self.feeds,
             events: self.events,
             radio_show: self.radio_show,
+            links: self.links,
+            membership_plans: self.membership_plans,
             issue_report_channels: self.issue_report_channels,
             state: self.state,
             extensions: self.extensions,
@@ -507,12 +573,74 @@ mod test {
     }
 
     #[test]
+    fn test_builder_v13_fail_on_location_timezone() {
+        let status = StatusBuilder::v0_13("foo")
+            .logo("bar")
+            .url("foobar")
+            .location(Location {
+                timezone: Some("Europe/London".into()),
+                ..Default::default()
+            })
+            .contact(Contact::default())
+            .add_issue_report_channel(IssueReportChannel::Email)
+            .state(State {
+                open: Some(false),
+                ..State::default()
+            })
+            .build();
+        assert!(status.is_err());
+        assert_eq!(
+            status.err().unwrap(),
+            "location.timezone is only present in v0.14 and above"
+        );
+    }
+
+    #[test]
+    fn test_builder_v13_fail_on_links() {
+        let status = StatusBuilder::v0_13("foo")
+            .logo("bar")
+            .url("foobar")
+            .contact(Contact::default())
+            .add_issue_report_channel(IssueReportChannel::Email)
+            .state(State {
+                open: Some(false),
+                ..State::default()
+            })
+            .add_link(Link::default())
+            .build();
+        assert!(status.is_err());
+        assert_eq!(status.err().unwrap(), "links is only present in v0.14 and above");
+    }
+
+    #[test]
+    fn test_builder_v13_fail_on_membership_plans() {
+        let status = StatusBuilder::v0_13("foo")
+            .logo("bar")
+            .url("foobar")
+            .contact(Contact::default())
+            .add_issue_report_channel(IssueReportChannel::Email)
+            .state(State {
+                open: Some(false),
+                ..State::default()
+            })
+            .add_membership_plan(MembershipPlan::default())
+            .build();
+        assert!(status.is_err());
+        assert_eq!(
+            status.err().unwrap(),
+            "membership_plans is only present in v0.14 and above"
+        );
+    }
+
+    #[test]
     fn test_builder_v14() {
         let status = StatusBuilder::v14("foo")
             .logo("bar")
             .url("foobar")
             .location(Location::default())
             .contact(Contact::default())
+            .add_link(Link::default())
+            .add_membership_plan(MembershipPlan::default())
             .build()
             .unwrap();
         assert_eq!(
@@ -524,6 +652,8 @@ mod test {
                 logo: "bar".into(),
                 url: "foobar".into(),
                 issue_report_channels: vec![],
+                links: Some(vec![Link::default()]),
+                membership_plans: Some(vec![MembershipPlan::default()]),
                 ..Status::default()
             }
         );